Add sf CLI check to the compile script
As part of the migration from the sfdx to sf CLIs, we want to add
additional output to the buildpack logs to show which version of the CLI
will be running in the build slug.

↔️: Using an OR list (`sf || sfdx`) is more elegant, but we want both
sets of output during our migration to assist in troubleshooting. We can
clean that up next time.